update inherit demo, add instance
