--- cpp_basics/inherit_demo/Demo.h
@@ -1,1 +1,0 @@
-#pragma once
